mqtt-gateway/gateway/requirements.txt
paho-mqtt
schedule
numpy<2.0.0
requests
